Begin streamlining Operating systems

+++ catscripts/Technology/Computers/Software/Operating_systems/OpenBSD/OpenBSD.sh
@@ -1,14 +1,21 @@
 #!/bin/bash
 
-egrep -i 'Open(| )BSD' newpages.txt >> OpenBSD.txt
+KEYWORDS_OPENBSD="Open(| )BSD"
 
-OPENBSD=`stat --print=%s OpenBSD.txt`
-
-if [ $OPENBSD -ne 0 ];
+if [ "$1" == "" ]; #Normal operation
 then
-  export CATFILE="OpenBSD.txt"
-  export CATNAME="OpenBSD"
-  $CATEGORIZE
-fi
 
-rm OpenBSD.txt
+  egrep -i "$KEYWORDS_OPENBSD" newpages.txt >> OpenBSD.txt
+
+  OPENBSD=`stat --print=%s OpenBSD.txt`
+
+  if [ $OPENBSD -ne 0 ];
+  then
+    export CATFILE="OpenBSD.txt"
+    export CATNAME="OpenBSD"
+    $CATEGORIZE
+  fi
+
+  rm OpenBSD.txt
+
+fi
